Fix generate_series type ambiguity in addRosters query
Cast parameters to ::int to resolve PostgreSQL function overload error.

+++ apps/backend/src/modules/leagues/league-rosters.repository.ts
@@ -20,7 +20,7 @@ export class LeagueRostersRepository {
     await this.db.query(
       `INSERT INTO rosters (roster_id, league_id)
        SELECT s, $1
-       FROM generate_series($2, $3) AS s`,
+       FROM generate_series($2::int, $3::int) AS s`,
       [leagueId, currentCount + 1, newCount]
     );
   }
